Add files via upload
I am presenting here the chat application that I have developed. I encountered some challenges during its development, which led to a delay. However, I have now implemented improvements.

The chat application consists of three main sections:

Login Screen: This is where users can log in the chat application using their user tokens. Users can log in  20 test users only , and if they make a mistake, shows  an error message will be displayed.


Chat List: This section displays a list of active users, active user is 20 testing  users. It shows the user numbers, their profile pictures, and indicates when they were last active. Unread messages are marked with a circle, and the latest message sent is displayed, similar to WhatsApp.

Chat View: In this section, shows where a two  chat conversation. It displays the number of messages in the chat, and so on the profile picture of the person being chatted with, and icons for making voice and video calls. The design is like  of the WhatsApp application.

I have put in significant effort, and I'm attaching some images for you to review. I'm hopeful for success, and I appreciate your ongoing support.  i wish successful inshallah